Cleaned up ErrorTests and renamed folder.
